refactor: use oreo's logger pkg
